requirements.txt
aiogram>=3.7.0
psycopg2-binary>=2.9.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytz>=2023.3
